BIG MOVING IN, all the v2 is now in the lexer directory. PTGLK
